Make typst-cli crate usable as library
